tests: move structures_views tests to insertion_views
